[devtools] [refactor] Move screen inspector to separate package
